Rebalanced the handheld drill and saw to have sensible damage values

+++ src/main/java/com/klarkson/creaternd/content/item/tool/handheldDrill/HandheldDrillItem.java
@@ -5,11 +5,9 @@ import com.simibubi.create.foundation.utility.VecHelper;
 import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.stats.Stats;
 import net.minecraft.tags.BlockTags;
-import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
@@ -34,15 +32,12 @@ import software.bernie.geckolib3.core.PlayState;
 import software.bernie.geckolib3.core.builder.AnimationBuilder;
 import software.bernie.geckolib3.core.builder.ILoopType;
 import software.bernie.geckolib3.core.controller.AnimationController;
-import software.bernie.geckolib3.core.event.predicate.AnimationEvent;
 import software.bernie.geckolib3.core.manager.AnimationData;
 import software.bernie.geckolib3.core.manager.AnimationFactory;
 import software.bernie.geckolib3.network.GeckoLibNetwork;
 import software.bernie.geckolib3.network.ISyncable;
 import software.bernie.geckolib3.util.GeckoLibUtil;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 import java.util.function.Consumer;
 
@@ -51,16 +46,14 @@ import java.util.function.Consumer;
 public class HandheldDrillItem extends PickaxeItem implements IAnimatable, ISyncable {
     private static final int ANGLE_LIMIT = 55;
 
-    private static final String SWINGING_BOOL = "swinging";
-    private static final String WAS_SWINGING_BOOL = "wasSwinging";
     private static final int ANIM_IDLE = 0;
     private static final int ANIM_DRILL = 1;
     private static final String CONTROLLER_NAME = "handheldDrillController";
-    public static final int MAX_DAMAGE = 360;
+    public static final int MAX_DAMAGE = 2048;
     public final AnimationFactory ANIMATION_FACTORY = GeckoLibUtil.createFactory(this);
 
-    public HandheldDrillItem(Tier tier, int alwaysOne, float attackSpeedBonus, Properties properties) {
-        super(tier, alwaysOne, attackSpeedBonus, properties.defaultDurability(MAX_DAMAGE));
+    public HandheldDrillItem(Tier tier, int attackBonus, float attackSpeedBonus, Properties properties) {
+        super(tier, attackBonus, attackSpeedBonus, properties.defaultDurability(MAX_DAMAGE));
         GeckoLibNetwork.registerSyncable(this);
     }
 
@@ -172,6 +165,9 @@ public class HandheldDrillItem extends PickaxeItem implements IAnimatable, ISync
                 }
 
                 BlockState blockMined = level.getBlockState(testPos);
+
+                if(blockMined.getTags().anyMatch(tag -> tag == BlockTags.NEEDS_DIAMOND_TOOL)) continue;
+
                 if(isCorrectToolForDrops(tool, blockMined)) {
                     tool.hurtAndBreak(1, player, (p) -> p.broadcastBreakEvent(EquipmentSlot.MAINHAND));
                     level.destroyBlock(testPos, false);
